System components can now be marked as readonly (const)
Passing a component type with a const modifier will notify the ECS that this component is readonly a given system
Meaning that the ECS can enforce constant access to every field of the component via the ComponentViews.
This will also notify the ECS update manager that a potential optimization can be done when updating systems, using multithreading, reducing frame times further along.