Systems can now signal failures while initializing (#26)
When a failure is notified by a system, the kernel will automatically stop any other initialization to exit with code 1000. If that system was non-critical (for example the logging system), a warning will be sent and initialization will continue.
In any case, a failed system won't be provided if requested.